Rename file upload method
